Fix bug causing AnimatedValue to freeze after same value is entered
There appears to be a bug in `Dashing.AnimatedValue` which causes the
AnimatedValue to freeze if the new value is equal to the current. The
source of the bug seems to be due to the comparison in `num == to`
being between a float and a string. By default, CoffeeScript compiles
`==` to Javascript's `===`. Instead, if we use backticks, we can use a
"raw Javascript" command which allows us to call `==`.