Remove websocket connections that have been closed in send_event
Currently, if a connection is not dropped gracefully its websocket stays
in the connection list. This means that when `send_event` runs it throws
an `IOError` when it gets to that connection, causing all subsequent
sockets in the list to not receive updated events. This commit instead
rescues the error and removes the closed websocket from the list of
connections.